Enable check of duplicate tasks
There is no check of duplicates when adding new tasks.

This causes issues when marking tasks as completed - it is not clear
whether they are the same task and accidentally duplicated or they are
multiple and distinct instances of the same task.

Adding this check would remove this ambiguity.

Let's add this check right before adding the new task to the list,
so that a response can be made to the user in lieu of a success
message.