Refactor parser
Validation and execution of all commands are done under one method.

This results in deep nesting, and a method that is longer than it
needs to be.

Pulling out the validation and execution from the method to form their
own distinct methods allows for greater readability of code.

Let's pull out the validation and execution of each command into their
separate methods, and call them under the original parsing method.

Doing so makes each component easier to debug as the different methods
can be tested separately.